# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SuperNØva Smart Locker API
  description: |
    API RESTful per il sistema di smart locker SuperNØva.

    ## Autenticazione
    L'API utilizza JWT (JSON Web Tokens) per l'autenticazione. Includi il token nell'header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 100 richieste per 15 minuti per IP
    - Risposta 429 quando il limite viene superato

    ## Errori
    Tutte le risposte di errore seguono il formato:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Descrizione dell'errore"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SuperNØva Support
    email: api@supernova-locker.it
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.supernova-locker.it/v1
    description: Production server
...
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: Lista tutti i dispositivi
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista dispositivi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
    post:
      tags: [Devices]
      summary: Registra nuovo dispositivo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Cassapanca Giardino
                serial_number:
                  type: string
                  example: CP2025-001
      responses:
        '201':
          description: Dispositivo registrato
  /devices/{deviceId}:
    get:
      tags: [Devices]
      summary: Dettagli dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Dettagli dispositivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Dispositivo non trovato
    patch:
      tags: [Devices]
      summary: Aggiorna dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                geofence:
                  $ref: '#/components/schemas/Geofence'
      responses:
        '200':
          description: Dispositivo aggiornato
    delete:
      tags: [Devices]
      summary: Elimina dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Dispositivo eliminato
  /devices/{deviceId}/lock:
    post:
      tags: [Devices]
      summary: Blocca dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Dispositivo bloccato
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      is_locked:
                        type: boolean
                        example: true
                      timestamp:
                        type: string
                        format: date-time
  /devices/{deviceId}/unlock:
    post:
      tags: [Devices]
      summary: Sblocca dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: Ritiro pacco
      responses:
        '200':
          description: Dispositivo sbloccato
  /devices/{deviceId}/status:
    get:
      tags: [Devices]
      summary: Stato corrente dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Stato dispositivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'

  # ============================================
  # Locations Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: ID univoco del dispositivo
      schema:
        type: string
        example: dev_cassapanca_001
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        model:
          type: string
        firmware_version:
          type: string
        hardware_specs:
          type: object
          properties:
            gps_module:
              type: string
            lte_module:
              type: string
            battery_capacity_mah:
              type: integer
            mcu:
              type: string
        current_location:
          $ref: '#/components/schemas/Location'
        status:
          $ref: '#/components/schemas/DeviceStatus'
        geofence:
          $ref: '#/components/schemas/Geofence'
        created_at:
          type: string
          format: date-time
    DeviceStatus:
      type: object
      properties:
        is_locked:
          type: boolean
        battery_percent:
          type: integer
        signal_strength:
          type: integer
        gps_fix:
          type: string
          enum: ['2D', '3D', 'none']
        last_checkin:
          type: string
          format: date-time
    Location:
      type: object
      properties:
        latitude:
          type: number
          example: 45.052
        longitude:
          type: number
          example: 9.695
        accuracy_m:
          type: number
          example: 2.5
        timestamp:
          type: string
          format: date-time
    Geofence:
      type: object
      properties:
        center:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        radius_m:
          type: number
          minimum: 10
          maximum: 10000
        alert_on_exit:
          type: boolean
          default: true
        alert_on_enter:
          type: boolean
          default: false
...